订单后台完成

+++ ss-manager/ss-manager-pojo/src/main/java/com/qf/pojo/TbMealOrder.java
@@ -3,9 +3,9 @@ package com.qf.pojo;
 public class TbMealOrder {
     private Long id;
 
-    private Integer mealId;
+    private Long mealId;
 
-    private Integer orderId;
+    private Long orderId;
 
     private Integer amount;
 
@@ -17,19 +17,19 @@ public class TbMealOrder {
         this.id = id;
     }
 
-    public Integer getMealId() {
+    public Long getMealId() {
         return mealId;
     }
 
-    public void setMealId(Integer mealId) {
+    public void setMealId(Long mealId) {
         this.mealId = mealId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
